move PluginLoadingException to internal.exception package
